Darn, with NVidia latest(*) I can't use Xv anymore for v4l devices. Worse, the mplayer frequency setting is not compatable with the Xv one. So make it default to MHz, and store it in kHz. (*)I could downgrade, but w/ xorg-7.1 video and composite is my personal    favorite disruptive technology :-)
svn path=/trunk/extragear/multimedia/kmplayer/; revision=580899

+++ src/kmplayertvsource.h
@@ -92,7 +92,7 @@ public:
  */
 class KMPLAYER_NO_EXPORT TVChannel : public TVNode {
 public:
-    TVChannel (KMPlayer::NodePtr & d, const QString & n, int f);
+    TVChannel (KMPlayer::NodePtr & d, const QString & n, double f);
     TVChannel (KMPlayer::NodePtr & d);
     KDE_NO_CDTOR_EXPORT ~TVChannel () {}
     void closed ();
